feat: device hook 內與 pin 有關的獨立成另一個檔案
